Auto-reset support for NodeMCU devkit as well as esprog,etc.
The NodeMCU dev-kit board has a similar auto-reset circuit to previous
ones (RTS drives reset, DTR drives GPIO0), but with a quirk - if both
RTS & DTR are asserted at once, neither nRESET nor GPIO0 are asserted.

This commit changes the reset order so RTS,DTR are asserted one at a
time instead of simultaneously. This should still work with the
traditional configuration as well.

This problem was mentioned in #26.