Better fix for CH341 slow baud rate bug
Based on information in #44
https://github.com/themadinventor/esptool/issues/44#issuecomment-107094446

Looks like if the kernels thinks the serial port is already set to
<baudrate>, it won't set the baud rate in set_termios. Due to the
pre-4.0 bug of always chosing 9600 baud on open, it's possible for the
device to be set to 9600 but the kernel thinks it's set much faster.

Not sure what I was thinking with the other fix in de97c54a0. It did
seem a bit more reliable but possibly this was because when testing I
kept changing the baud rate(!)