Wrap SLIP packet reads into a Python generator
With this change read_flash uses 7% of the read/select
syscalls. read_flash spends about 50% less time in syscalls on my i7,
which should improve performance on slow systems like Raspberry Pi.

A side effect of the change is an unintentional fix to sync() behaviour,
which didn't retry cleanly before. Less of a problem now that we sync at
max 115200bps, though.