serial port: Set timeout before writing command, only if changed
Fixes some bugs on macOS where setting timeout in pyserial calls
_reconfigure_port() which sends a bunch of config settings to the
port (unnecessarilly as timeout is implemented elsewhere on POSIX). This seems
to cause the input buffer to flush in some circumstances, so the subsequent
read() times out.

On macOS, the bug that turned this up was flashing failing at
>230400bps. Unclear if this is because the response is received in less time, or
because >230400bps uses different ioctls() to set.

This may fix some of the following issues:
https://github.com/espressif/esptool/issues/250
https://github.com/espressif/esptool/issues/255
https://github.com/espressif/esptool/pull/247