esptool, espefuse: Add new argument "--before no_sync_no_reset"
We have a board designed which has the UART to the ESP32 as a passthrough from
another microcontroller.  As such, it's easier to send the sync sequence and
change the baud rate on the microcontroller side of things than to pass this
change through from the PC.  Adding this new option allows the intermediate
microcontroller to run the sync sequence and to set the appropriate baud rate
on the ESP32 side of things.  For this use case, we need to avoid sending the
change baud rate sequence (the PC baud rate is not necessarily the same as
the microcontroller <-> ESP32) and avoid sending the sync sequence.

Signed-off-by: Tim Nordell <tim.nordell@nimbelink.com>