elf2image: add workaround for mapping bug in IDF bootloader
ESP-IDF 2nd stage bootloader had a bug that it mapped one page less
than needed, if an IROM/DROM segment was 0x20 or less bytes over a
64kB MMU page boundary. This is fixed in IDF as well, but to prevent
having the issue with old bootloaders already deployed, adding a
workaround into esptool. The workaround is to pad the segment to 0x24
bytes over the MMU page boundary.