flasher_stub: Change the loading location to facilitate loading of applications that use SRAM1 as IRAM
The flasher stub text is moved to end of SRAM1 and the data is moved to
end of DRAM (start of SRAM2 with inverted addressing). This will allow
loading of the application in memory with maximum IRAM and data
sections. This text and data sections will get added to heap area when
the application executes.

Signed-off-by: Amey Inamdar <amey@espressif.com>