Made Arduino IDE scroll during upload
When esptool is run from the Arduino IDE, the output window doesn't
scroll because of the \rs used to make the console update status on a
single line.

To avoid this, don't attempt to output everything on a single line when
scripted (i.e. not running on a TTY) by replacing \r with \n.

See https://github.com/esp8266/Arduino/pull/6765 for some discussion.