Update miniz to 2.1.0
Update miniz to the latest version, 2.1.0. This is the stock, amalgated
miniz, with a few modifications, forward-ported from the existing diff
against 1.5:
 * Configuration options (the definitions at the top of miniz.h)
 * Disabling assert(), i.e. defining MZ_ASSERT to nothing
 * Defining TDEFL_LESS_MEMORY to 1
 * Setting TDEFL_LZ_DICT_SIZE = 8*1024 and TDEFL_LZ_CODE_BUF_SIZE = 8 * 1024

An additional complication is that the new miniz performs a long long to
integer division, which requires code from libgcc (__ashldi3), which we
didn't link to by default (due to -nostdlib). Add -lgcc to LDLIBS.

This seems to be a small increase in .text (8018->8023) with the stock
toolchain, but with a newer one (gcc 9.2.1/picolibc 1.3) actually
results in a decrease (7990->7947).

Tested on real hardware (D1 Mini Pro).

See #499 for some additional context.