fix(elf2image): fix elf2image for ram app when sha256 offset not specified
This commit fixes issue where elf2image fails to generate image for ram
app when sha256 offset is not specified. This commit adds a check to
verify if the app description segment is in the flash region and if not
it will not add the sha256 offset to the image header.