fix: enable ESP32-P4 ECO5 chip detection
Register was read in order to detect SDM, this cannot be done with new
ESP32-P4 ECO5, because it has different address space and crashes
when trying to read current magic address. This also improves detection
for ESP32-S2 as is is only chip that has SDM support and is detected
by reading magic address.